before codegen
