+++ skyrim-platform/src/platform_se/skyrim_platform/NirnLabCodegen.cpp
@@ -1,0 +1,7 @@
+#include "NirnLabCodegen.h"
+#include "FunctionsDumpFormat.h"
+
+std::string NirnLabCodegen::GenerateCode(const FunctionsDumpFormat::Root& root)
+{
+  return "";
+}
