+++ skyrim-platform/src/platform_se/skyrim_platform/NirnLabCodegen.h
@@ -1,0 +1,11 @@
+#pragma once
+#include <string>
+
+namespace FunctionsDumpFormat {
+    struct Root;
+}
+
+class NirnLabCodegen {
+public:
+  static std::string GenerateCode(const FunctionsDumpFormat::Root& root);
+};
